+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New public attributes for the top, bottom and left margins (#8).
+- New component classes to style the axes, and the tick and axis labels.
 
 ## [0.7.0] - 2025-07-14